+++ app/functions/growth.py
@@ -9,6 +9,8 @@ from app.functions import maths
 
 def calculate_growth(c,t,params):
     # Get data regarding growth
+    messages = []
+    logging.info(f"EXOADMIN: health requirement {params.get('pop_health_requirement')}")
     healthy_pops_query = f"""
             g.V().has('label','pop')
                 .has('health',gt({params.get("pop_health_requirement")})).as('pop')
@@ -17,7 +19,10 @@ def calculate_growth(c,t,params):
     c.run_query(healthy_pops_query)
 
     a = np.array(c.res)
-    np.split(a,len(a)/3)
+    logging.info(f"EXOADMIN: healthy_pops_query {len(a)}")
+    if len(a)<=0:
+        logging.info(f"EXOADMIN: No pops that meet the pop_health_requirement")
+        return messages
     pops_df = pd.DataFrame(np.split(a,len(a)/3),columns=['objid','health','wealth'])
     pops_df[['health','wealth']] = pops_df[['health','wealth']].astype(float)
 
@@ -27,20 +32,42 @@ def calculate_growth(c,t,params):
     reproducing_pops = pops_df[pops_df['grow']].drop(['roll','grow'],axis=1).reset_index(drop=True)
 
 
-    messages = []
     if len(reproducing_pops)==0:
-        logging.info(f'**** No pops capable of reproducing ****')
-        logging.info(f"**** Growth Complete *****")
+        logging.info(f'EXOADMIN: **** No pops capable of reproducing ****')
+        logging.info(f"EXOADMIN: **** Growth Complete *****")
         return messages
     
     if len(reproducing_pops)>0:
         logging.info(f"{len(reproducing_pops)} of {len(pops_df)} pops will grow")
 
         for parent_pop in reproducing_pops.to_dict(orient='records'):
-            logging.info(f"Growing Pop {parent_pop['objid']}")
-            messages.append(get_growth_message(parent_pop))
-        logging.info(f"**** Growth Complete *****")
+            can_grow = check_pop_growth(c, parent_pop)
+            if can_grow:
+                logging.info(f"EXOADMIN: Growing Pop {parent_pop['objid']}")
+                messages.append(get_growth_message(parent_pop))
+            if not can_grow:
+                logging.info(f"EXOADMIN: Pop {parent_pop['objid']} cannot grow")
+        logging.info(f"EXOADMIN: **** Growth Complete *****")
         return messages
+
+
+def check_pop_growth(c, parent_pop):
+    query_pop_cap = f"""
+        g.V().has('objid','{parent_pop["objid"]}').out('inhabits').values('objid','name','pop_cap')"
+        """
+    c.run_query(query_pop_cap)
+    pop_cap = c.res
+    query_location_pop = f"""
+        g.V().has('objid','{parent_pop["objid"]}').out('inhabits').in('inhabits').has('label','pop').count()"
+        """
+    c.run_query(query_location_pop)
+    location_pop = c.res
+    over_pop = location_pop[0] > pop_cap[2]
+    if not over_pop:
+        return True
+    else:
+        return False
+
     
 
 def get_growth_message(pop):
@@ -84,23 +111,41 @@ def grow_population(c,t, parent_pop):
     
     inhabits_edge = {'node1': child.objid, 'node2': loc_dict['objid'], 'label': 'inhabits'}
 
-    event = population_growth_event(t,pop_dict,loc_dict,child)
-    event_edge = {'node1': child.objid, 'node2': event['objid'], 'label': 'caused'}
-
-    data = {"nodes":[child_data, event] ,"edges":[inhabits_edge,child.isOfSpecies,isIn_edge,event_edge] }
+    data = {"nodes":[child_data] ,"edges":[inhabits_edge,child.isOfSpecies,isIn_edge] }
     c.upload_data(pop_dict['userguid'],data)
     return data
 
 
-def population_growth_event(t,parent,location,child):
-    node = {
-        'objid':maths.uuid(),
-        'name':'population growth',
-        'label':'event',
-        'text': f"The population ({parent['name']}) inhabiting {location['name']} has grown to produce the population: {child.name}.",
-        'visibleTo':parent['userguid'],
-        'time':t.params['currentTime'],
-        'userguid':parent['userguid'],
-        'source':'notebook'
-    }
-    return node
+def get_renewal_message(resource):
+    message = {"agent":resource,"action":"renew"}
+    return message
+
+def calculate_renewal(c,t,params):
+    renewing_resources_query =f"""
+    g.V().has('label','resource').has('replenish_rate').valuemap()
+    """
+    c.run_query(renewing_resources_query)
+    renewing_resources = c.clean_nodes(c.res)
+    messages = []
+    for resource in renewing_resources:
+        if float(resource['volume']) < float(resource['max_volume']):
+            messages.append(get_renewal_message(resource))
+    return messages
+
+def renew_resource(c,message):
+    objid = message['agent']['objid']
+    old_volume = float(message['agent']['volume'])
+    replenish_rate = float(message['agent']['replenish_rate'])
+    max_volume = float(message['agent']['max_volume'])  
+
+    new_volume = old_volume + replenish_rate
+    if new_volume > max_volume:
+        new_volume = max_volume
+
+    patch_resource_query = f"""
+    g.V().has('objid','{objid}').has('label','resource')
+        .property('volume', {new_volume})
+    """
+    c.run_query(patch_resource_query)
+    logging.info(f"EXOADMIN: {message['agent']['name']}:{message['agent']['objid']} increased by {replenish_rate}, {old_volume}-> {new_volume}")
+    return None
